+++ DeepNtuplizer/interface/ntuple_pfCands.h
@@ -31,7 +31,7 @@ private:
 	float  Cpfcan_phirel_[100];
 	float  Cpfcan_etarel_[100];
 	float  Cpfcan_puppiw_[100];
-	int   Cpfcan_VTX_ass_[100];
+	float   Cpfcan_VTX_ass_[100];
 
 	// covariance
 	float  Cpfcan_dz_[100];
@@ -47,21 +47,21 @@ private:
 
 	// ID, skipped "charged hadron" as that is true if now the other
 	// TODO (comment of Markus Stoye) add reco information
-	int Cpfcan_isMu_[100]; // pitty that the quality is missing
-	int Cpfcan_isEl_[100]; // pitty that the quality is missing
-	int Cpfcan_charge_[100];
+	float Cpfcan_isMu_[100]; // pitty that the quality is missing
+	float Cpfcan_isEl_[100]; // pitty that the quality is missing
+	float Cpfcan_charge_[100];
 
 	// track quality
-	int Cpfcan_lostInnerHits_[100];
+	float Cpfcan_lostInnerHits_[100];
 	float Cpfcan_chi2_[100];
-	int Cpfcan_highPurity_[100];
+	float Cpfcan_highPurity_[100];
 
 	//Neutral Pf candidates
 	int n_Npfcand_;
 	float  Npfcan_pt_[100];
 	float  Npfcan_phirel_[100];
 	float  Npfcan_etarel_[100];
-	int  Npfcan_isGamma_[100];
+	float  Npfcan_isGamma_[100];
 	float  Npfcan_HadFrac_[100];
 
 };
